fixed code for vtk 8.2 build with tk support
User will need to link python to vtk8's cmake configuration and test if there are more build problems. For me, if it has to involve python, I'd go for vtk and qt in python. Not that I'm against python, but this repo is specifically designed for tcl/tk and vtk integration alone.